backend/requirements.txt
django==5.2.10
djangorestframework==3.16.1
django-filter==25.2
djoser==2.3.3
pillow==12.1.0
python-dotenv==1.2.1
gunicorn==23.0.0
ruff==0.14.13
pre-commit==4.5.1
pillow==12.1.0
weasyprint==68.0
psycopg==3.3.2
